==> packages/kptn/kptn-0.1.4.tar.gz/kptn-0.1.4/example/nibrs/tasks/generate_estimates/Tables_Core/Table2a/Part3_finalize.R <==
library(tidyverse)
library(openxlsx)
library(DBI)
library(DT)
library(lubridate)
library(data.table)

source(here::here("tasks/logging.R"))
source(here::here("tasks/impute_items/0-Common_functions_for_imputation.R"))

source("util.R")

DER_TABLE_NAME <- Sys.getenv("DER_TABLE_NAME")
inputPipelineDir <- Sys.getenv("INPUT_PIPELINE_DIR")
outputPipelineDir <- Sys.getenv("OUTPUT_PIPELINE_DIR")
in_file_path = paste0(inputPipelineDir, "/indicator_table_single_intermediate/")
load(paste0(in_file_path,"/",DER_TABLE_NAME,"_prep_env.RData"))

#Figure out the number of columns
der_determine_column <- list.files(path=in_file_path, pattern=paste0("data_", DER_TABLE_NAME, "_")) %>%
  as_tibble() %>%
  mutate(der_table_num = str_match(string=.$value, pattern="\\w+_\\w+_(\\d+)\\.rds")[,2] %>% as.numeric()) %>%
  summarise(max_column = max(der_table_num)) %>%
  pull()

#Print the column
der_determine_column

#Overwrite maximum column
maximum_column <- der_determine_column


log_info("Running 05_Table2a_part3_finalize.R")
log_debug(system("free -mh", intern = FALSE))

#Create the filler dataset
data0 <- c(1:DER_MAXIMUM_ROW) %>% as_tibble() %>%
  rename(row = value)
data0 <- assign_section(data0)

#Create the list objects to separate
final_reporting_database <- vector("list", maximum_column)
final_data_list <- vector("list", maximum_column+1)
final_ori_data <- vector("list", maximum_column+1)

#Add on the filler dataset
final_data_list[[1]] <- data0
final_ori_data[[1]] <- weight_dataset

#Separate out the data from list objects
for(i in 1:maximum_column){

  #Get the data object
  raw_object <- readRDS(paste0(out_file_path,"/data_",DER_TABLE_NAME,"_",i,".rds"))


  #Save normal table to list
  final_data_list[[i+1]] <- raw_object[[1]]

  #Save the reporting database to new object
  final_reporting_database[[i]] <-   raw_object[[2]]

  final_ori_data[[i+1]] <-   raw_object[[3]]

  #Delete the raw objects
  rm(raw_object)
  invisible(gc())

}

#Clear up memory
rm(list = c(ls(pattern="^agg"), ls(pattern="^data\\d+"), ls(pattern="^main")))
invisible(gc())


#Output the Reporting database
final_reporting_database2 <- final_reporting_database %>%
  bind_rows()

#Choose columns to not be transpose
final_reporting_database2_column <- final_reporting_database2 %>%
  head(1) %>%
  colnames() %>%
  as_tibble() %>%
  filter(!trim_upcase(value) %in% c("COUNT", "PERCENTAGE", "RATE") ) %>%
  select(1) %>%
  pull() %>%
  paste0("-", .) %>%
  rlang:::parse_exprs()

#Create the final database
final_reporting_database2 %>%
  gather(key="estimate_type", value="estimate", !!!final_reporting_database2_column) %>%
  #Clear population_estimate if not estimate_type RATE
  mutate(population_estimate = case_when(!trim_upcase(estimate_type) %in% c("RATE") ~ DER_NA_CODE,
                                         TRUE ~ population_estimate),
         estimate_geographic_location = DER_GEOGRAPHIC_LOCATION,
         analysis_weight_name = DER_WEIGHT_VARIABLE_STRING,
         estimate_type_num = case_when(
            trim_upcase(estimate_type) %in% c("COUNT") ~ 1,
            trim_upcase(estimate_type) %in% c("PERCENTAGE") ~ 2,
            trim_upcase(estimate_type) %in% c("RATE") ~ 3
         )) %>%
  #Create the estimate_type_detail_percentage variable
  estimate_type_detail_percentage_label() %>%
  #Create the estimate_type_detail_rate variable
  estimate_type_detail_rate_label() %>%
  write.csv(paste0(final_path,"Table ", DER_TABLE_NAME, "_Reporting_Database.csv"))

#Output the ORI domains for Variance

#Create the final ORI dataset
final_ori_data2 <- final_ori_data %>%
  reduce(left_join, by = c("ori")) %>%
  write.csv0(gzfile(paste0(final_path,"Table ", DER_TABLE_NAME, " ORI_PERSON.csv.gz")))

log_debug("Ending table")
log_debug(system("free -mh", intern = FALSE))
